requirements.txt
requests==2.26.0
vcrpy==4.1.1
pytest==6.2.5